@@ -1,8 +1,8 @@
 {
   "name": "bnomei/kirby3-seobility",
-  "description": "Kirby 3 Plugin to use Seobility.net",
+  "description": "Kirby Plugin to use Seobility.net",
   "type": "kirby-plugin",
-  "version": "1.4.1",
+  "version": "5.0.0",
   "license": "MIT",
   "authors": [
     {
@@ -11,10 +11,11 @@
     }
   ],
   "keywords": [
-    "kirby3",
-    "kirby3-cms",
-    "kirby3-plugin",
+    "kirby",
+    "kirby-cms",
+    "kirby-plugin",
     "seo",
+    "seobility",
     "keyword",
     "check",
     "site",
@@ -25,36 +26,36 @@
     "suggestion",
     "ranking"
   ],
+  "config": {
+    "optimize-autoloader": true,
+    "sort-packages": true,
+    "allow-plugins": {
+      "getkirby/composer-installer": true,
+      "pestphp/pest-plugin": true
+    }
+  },
   "autoload": {
     "psr-4": {
       "Bnomei\\": "classes/"
     }
   },
-  "config": {
-    "optimize-autoloader": true,
-    "sort-packages": true,
-    "allow-plugins": {
-      "getkirby/composer-installer": true
-    }
-  },
   "require": {
-    "php": ">=7.4.0",
+    "php": ">=8.2",
     "getkirby/composer-installer": "^1.2"
   },
   "require-dev": {
-    "getkirby/cms": "^3.6",
-    "php-coveralls/php-coveralls": "^2.1",
-    "phpunit/phpunit": "^9.5"
+    "getkirby/cms": "^5.0.0-alpha.4",
+    "larastan/larastan": "^v3.0.0",
+    "laravel/pint": "^1.13",
+    "pestphp/pest": "^v3.5.1",
+    "spatie/ray": "^1.39"
   },
   "scripts": {
-    "analyze": "phpstan analyse classes",
-    "fix": "php-cs-fixer fix",
-    "test": [
-      "mkdir -p tests/logs",
-      "@putenv XDEBUG_MODE=coverage",
-      "phpunit --configuration ./phpunit.xml"
-    ],
+    "stan": "./vendor/bin/phpstan",
+    "fix": "./vendor/bin/pint",
+    "test": "./vendor/bin/pest --profile",
     "dist": [
+      "composer fix",
       "composer install --no-dev --optimize-autoloader",
       "git rm -rf --cached .; git add .;"
     ],
@@ -62,7 +63,9 @@
       "composer install",
       "composer update",
       "composer install --working-dir=tests/kirby --no-dev --optimize-autoloader",
-      "composer update --working-dir=tests/kirby"
+      "composer update --working-dir=tests/kirby",
+      "sed -i.bak 's/function dump(/function xdump(/g' tests/kirby/config/helpers.php",
+      "sed -i.bak 's/function e(/function xe(/g' tests/kirby/config/helpers.php"
     ]
   },
   "extra": {
