@@ -4,30 +4,30 @@ declare(strict_types=1);
 
 namespace Bnomei;
 
+use Closure;
 use Exception;
+use Kirby\Cms\Page;
 use Kirby\Data\Json;
 use Kirby\Http\Remote;
 use Kirby\Toolkit\A;
 
 final class Seobility
 {
-    /** @var array $options */
-    private $options;
+    private array $options;
 
     public function __construct(array $options = [])
     {
-        $defaults = [
+        $this->options = array_merge([
             'debug' => option('debug'),
-            'version' => Json::read(__DIR__ . '/../composer.json')['version'],
+            'version' => Json::read(__DIR__.'/../composer.json')['version'],
             'enabled' => option('bnomei.seobility.enabled'),
             'expire' => intval(option('bnomei.seobility.expire')),
             'apikey' => option('bnomei.seobility.apikey'),
             'searchengine' => option('bnomei.seobility.searchengine'),
-        ];
-        $this->options = array_merge($defaults, $options);
+        ], $options);
 
         foreach ($this->options as $key => $call) {
-            if (is_callable($call) && in_array($key, ['apikey', 'searchengine'])) {
+            if ($call instanceof Closure && in_array($key, ['apikey', 'searchengine'])) {
                 $this->options[$key] = $call();
             }
         }
@@ -45,15 +45,12 @@ final class Seobility
         }
     }
 
-    /**
-     * @param string|null $key
-     * @return array|mixed
-     */
-    public function option(?string $key = null)
+    public function option(?string $key = null): mixed
     {
         if ($key) {
             return A::get($this->options, $key);
         }
+
         return $this->options;
     }
 
@@ -73,55 +70,56 @@ final class Seobility
         } else {
             kirby()->cache('bnomei.seobility')->set('Error', 1); // minutes
         }
+
         return null;
     }
 
-    public function keywordcheck(\Kirby\Cms\Page $page, ?string $keyword = null, ?string $url = null): array
+    public function keywordcheck(Page $page, ?string $keyword = null, ?string $url = null): array
     {
         $check = $url ?? $page->url();
-        $keyword = $keyword ?? $page->keywordcheck()->value();
+        $keyword = $keyword ?? $page->keywordcheck()->value(); // @phpstan-ignore-line
 
         $default = [
             'modified' => $page->modified(),
             'score' => 0,
             // 'keyword' => $keyword,
-            'url' => option('bnomei.seobility.free.keywordcheck')($check, $keyword),
+            'url' => option('bnomei.seobility.free.keywordcheck')($check, $keyword), // @phpstan-ignore-line
         ];
 
-        if (!$this->option('enabled') || $page->isDraft() ||
+        if (! $this->option('enabled') || $page->isDraft() ||
             empty($keyword) || kirby()->cache('bnomei.seobility')->get('Error')) {
             return $default;
         }
 
-        $key = md5('keywordcheck' . $page->url() . $keyword . $this->option('apikey'));
+        $key = md5('keywordcheck'.$page->url().$keyword.$this->option('apikey'));
 
         $data = kirby()->cache('bnomei.seobility')->get($key);
-        if (!$data || intval(A::get($data, 'modified')) < $page->modified()) {
+        if (! $data || intval(A::get($data, 'modified')) < $page->modified()) {
             $data = $default;
-            if ($this->option('apikey') && $remote = $this->api(option('bnomei.seobility.paid.keywordcheck')($check, $keyword))) {
+            if ($this->option('apikey') && $remote = $this->api(option('bnomei.seobility.paid.keywordcheck')($check, $keyword))) { // @phpstan-ignore-line
                 $data['score'] = intval(A::get($remote->json(), 'score', 0));
-            } elseif ($remote = $this->api(option('bnomei.seobility.free.keywordcheck')($check, $keyword))) {
+            } elseif ($remote = $this->api(option('bnomei.seobility.free.keywordcheck')($check, $keyword))) { // @phpstan-ignore-line
                 // scrapper
                 preg_match_all(
                     '/"data":(.*?),/',
-                    (string)$remote->content(),
+                    (string) $remote->content(),
                     $matches
                 );
-                if ($matches && count($matches[1])) {
+                if (count($matches[1])) {
                     $data['score'] = intval($matches[1][0]);
                 }
             }
 
-            kirby()->cache('bnomei.seobility')->set($key, $data, $this->option('expire'));
+            kirby()->cache('bnomei.seobility')->set($key, $data, intval($this->option('expire')));
         }
 
         return $data;
     }
 
-    public function ranking(\Kirby\Cms\Page $page, ?string $keyword = null, ?string $url = null): array
+    public function ranking(Page $page, ?string $keyword = null, ?string $url = null): array
     {
         $check = $url ?? $page->url();
-        $keyword = $keyword ?? $page->keywordcheck()->value();
+        $keyword = $keyword ?? $page->keywordcheck()->value(); // @phpstan-ignore-line
 
         $default = [
             'modified' => $page->modified(),
@@ -129,18 +127,18 @@ final class Seobility
             // 'keyword' => $keyword,
         ];
 
-        if (!$this->option('enabled') || empty($this->option('apikey')) ||
+        if (! $this->option('enabled') || empty($this->option('apikey')) ||
             $page->isDraft() || empty($keyword) ||
             kirby()->cache('bnomei.seobility')->get('Error')) {
             return $default;
         }
 
-        $key = md5('ranking' . $page->url() . $keyword . $this->option('apikey'));
+        $key = md5('ranking'.$page->url().$keyword.$this->option('apikey'));
 
         $data = kirby()->cache('bnomei.seobility')->get($key);
-        if (!$data || intval(A::get($data, 'modified')) < $page->modified()) {
+        if (! $data || intval(A::get($data, 'modified')) < $page->modified()) {
             $data = $default;
-            if ($remote = $this->api(option('bnomei.seobility.paid.ranking')($check, $keyword))) {
+            if ($remote = $this->api(option('bnomei.seobility.paid.ranking')($check, $keyword))) { // @phpstan-ignore-line
                 if ($targeturlresult = A::get($remote->json(), 'targeturlresult')) {
                     $data['rank'] = intval(A::get($targeturlresult, 'rank'));
                     $data['title'] = A::get($targeturlresult, 'title');
@@ -148,16 +146,16 @@ final class Seobility
                 }
             }
 
-            kirby()->cache('bnomei.seobility')->set($key, $data, $this->option('expire'));
+            kirby()->cache('bnomei.seobility')->set($key, $data, intval($this->option('expire')));
         }
 
         return $data;
     }
 
-    public function termsuggestion(\Kirby\Cms\Page $page, ?string $keyword = null, ?string $url = null): array
+    public function termsuggestion(Page $page, ?string $keyword = null, ?string $url = null): array
     {
         $check = $url ?? $page->url();
-        $keyword = $keyword ?? $page->keywordcheck()->value();
+        $keyword = $keyword ?? $page->keywordcheck()->value(); // @phpstan-ignore-line
 
         $default = [
             'modified' => $page->modified(),
@@ -167,18 +165,18 @@ final class Seobility
             // 'keyword' => $keyword,
         ];
 
-        if (!$this->option('enabled') || empty($this->option('apikey')) ||
+        if (! $this->option('enabled') || empty($this->option('apikey')) ||
             $page->isDraft() || empty($keyword) ||
             kirby()->cache('bnomei.seobility')->get('Error')) {
             return $default;
         }
 
-        $key = md5('termsuggestion' . $page->url() . $keyword . $this->option('apikey'));
+        $key = md5('termsuggestion'.$page->url().$keyword.$this->option('apikey'));
 
         $data = kirby()->cache('bnomei.seobility')->get($key);
-        if (!$data || intval(A::get($data, 'modified')) < $page->modified()) {
+        if (! $data || intval(A::get($data, 'modified')) < $page->modified()) {
             $data = $default;
-            if ($remote = $this->api(option('bnomei.seobility.paid.termsuggestion')($check, $keyword))) {
+            if ($remote = $this->api(option('bnomei.seobility.paid.termsuggestion')($check, $keyword))) { // @phpstan-ignore-line
                 if ($termsuggestions = A::get($remote->json(), 'termsuggestions')) {
                     $data['more'] = A::get($termsuggestions, 'more');
                     $data['less'] = A::get($termsuggestions, 'less');
@@ -187,22 +185,17 @@ final class Seobility
                 }
             }
 
-            kirby()->cache('bnomei.seobility')->set($key, $data, $this->option('expire'));
+            kirby()->cache('bnomei.seobility')->set($key, $data, intval($this->option('expire')));
         }
 
         return $data;
     }
 
-    /** @var Seobility */
-    private static $singleton;
+    private static ?self $singleton = null;
 
-    /**
-     * @param array $options
-     * @return Seobility
-     */
-    public static function singleton(array $options = [])
+    public static function singleton(array $options = []): self
     {
-        if (!self::$singleton) {
+        if (self::$singleton === null) {
             self::$singleton = new self($options);
         }
 
