@@ -5,7 +5,7 @@ return [
     'default' => false,
     'direction' => 'ltr',
     'locale' => [
-        'LC_ALL' => 'de_DE'
+        'LC_ALL' => 'de_DE',
     ],
     'name' => 'Deutsch',
     'translations' => [],
